alerter/requirements.txt
celery
psycopg2-binary
requests
redis